refactor: add package for recyclerview item data classes(model.rv_item)
