+++ app/src/main/java/com/example/yomakase/adapter/ClosedDayAdapter.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.example.yomakase.model.ClosedDay
+import com.example.yomakase.model.rv_item.ClosedDay
 import com.example.yomakase.databinding.RvItemClosedDayBinding
 
 class ClosedDayAdapter(private val closedDayList:List<ClosedDay>, val onClickItemListener: (closedDay: View, position: Int) -> Unit)
