+++ app/src/main/java/com/example/yomakase/adapter/DlgFacilityAdapter.kt
@@ -3,7 +3,7 @@ package com.example.yomakase.adapter
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.example.yomakase.model.DialogFacility
+import com.example.yomakase.model.rv_item.DialogFacility
 import com.example.yomakase.databinding.RvItemDlgFacilityBinding
 
 class DlgFacilityAdapter(private val itemList: List<DialogFacility>
