+++ app/src/main/java/com/example/yomakase/adapter/PriceAdapter.kt
@@ -3,7 +3,7 @@ package com.example.yomakase.adapter
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.example.yomakase.model.Price
+import com.example.yomakase.model.rv_item.Price
 import com.example.yomakase.databinding.RvItemPriceBinding
 
 class PriceAdapter(private val priceList:List<Price>, val onClickRemoveBtn: (price: Price) -> Unit): RecyclerView.Adapter<PriceAdapter.ViewHolder>() {
