+++ app/src/main/java/com/example/yomakase/dialog/FacilityDialog.kt
@@ -11,7 +11,7 @@ import androidx.fragment.app.DialogFragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.yomakase.adapter.DlgFacilityAdapter
 import com.example.yomakase.databinding.DlgFacilitiesBinding
-import com.example.yomakase.model.DialogFacility
+import com.example.yomakase.model.rv_item.DialogFacility
 
 class FacilityDialog(facilityDialogInterface: FacilityDialogInterface, context: Context,  var preSelected: List<DialogFacility>): DialogFragment(){
     private var _binding: DlgFacilitiesBinding? = null
