+++ app/src/main/java/com/example/yomakase/model/rv_item/ClosedDay.kt
@@ -1,4 +1,4 @@
-package com.example.yomakase.model
+package com.example.yomakase.model.rv_item
 
 data class ClosedDay(
     val day: String
