+++ app/src/main/res/layout/rv_item_closed_day.xml
@@ -4,7 +4,7 @@
     <data>
         <variable
             name="closeday"
-            type="com.example.yomakase.model.ClosedDay" />
+            type="com.example.yomakase.model.rv_item.ClosedDay" />
     </data>
 
     <LinearLayout
