+++ app/src/main/res/layout/rv_item_dlg_facility.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="facility"
-            type="com.example.yomakase.model.DialogFacility" />
+            type="com.example.yomakase.model.rv_item.DialogFacility" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
