+++ app/src/main/res/layout/rv_item_price.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="price"
-            type="com.example.yomakase.model.Price" />
+            type="com.example.yomakase.model.rv_item.Price" />
     </data>
 
     <LinearLayout
